[iOS] refactor: MainViewLinkManager 파일 분리
