+++ iOS/DaDaIkSeon/DaDaIkSeon/Scene/Main/View/MainView.swift
@@ -8,41 +8,6 @@
 import SwiftUI
 import UniformTypeIdentifiers
 
-class MainLinkManager: ObservableObject {
-    
-    @Published var tag: Int? = nil
-    
-    func isThere(_ target: MainLinkTable) -> Bool {
-        tag == scene(target)
-    }
-    
-    func change(_ target: MainLinkTable) {
-        DispatchQueue.main.async { [weak self] in
-            guard let self = self else { return }
-            self.tag = self.scene(target)
-        }
-    }
-    
-    func scene(_ target: MainLinkTable) -> Int? {
-        switch target {
-        case .main:
-            return nil
-        case .qrguide, .tokenEdit,
-             .background, .pincode, .setting:
-            return target.rawValue
-        }
-    }
-    
-    enum MainLinkTable: Int {
-        case main = -1
-        case qrguide = 0
-        case tokenEdit = 1
-        case background = 2
-        case pincode = 3
-        case setting = 4
-    }
-}
-
 struct MainView: View {
     
     // MARK: ViewModel
